requirements.txt
dateutils==0.6.12
loguru==0.7.2
pod2gen==1.0.3
pycountry==24.6.1
pydantic==2.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
